requirements.txt
streamlit>=1.24.0
pandas>=2.0.0
httpx>=0.24.0
requests>=2.28.0
python-dotenv>=1.0.0
gspread>=5.10.0
oauth2client>=4.1.3
openai>=1.0.0
plotly>=5.15.0
pytest>=7.4.0 